Porting using //eve/development/MAINLINE<-//eve/staging/2014-DOLLY at head (801633)
========================================

Change 798686 by brianb@games_brianb_integration_funtimes on 2014/06/13 15:32:30

Accounts are being marked as online after failing to login due to VIP - problem for trial accounts
https://jira.ccpgames.com/browse/EVE-111139

Since we have to Auth in people before checking VIP now that we're using the SSO, we need to log them out if they fail VIP.  This means a standard using logging in during VIP hits the DB twice which is non-ideal but it's a needed step until we have VIP handled by the SSO.

========================================

Change 800363 by stevem@STEVEM-WS on 2014/06/16 11:07:39

DEFECT: CluError: eve/client/script/parklife\spaceMgr.py(152) LoadObject RuntimeError: Couldn't get object
https://jira.ccpgames.com/browse/EVE-113060

The Sansha Fighter Bomber is made for a drone usage, not for NPC usage. This caused this error. I caught the cause of this error better, but really "fixing" it is a feature: "Make drones usable by NPCs"

Testing: give missiles, both from players and NPCs, a good testing

========================================

Change 800738 by keke@KEKE-PC on 2014/06/17 07:22:31

No Jira link

- Backout JoinCorpMembers hack for fanfest
- The hack change was 775639

Reviewed by: max

========================================

Change 801633 by stephen@stephen_snorthcott-ws_evegl on 2014/06/18 10:43:51

While investigating the following defect : "Texture bug at recustomization" ( https://jira.ccpgames.com/browse/EVE-113178 ), Bara discovered that the function BakeBlendshapeFromScript was not returning its error state to the Python function calling it - which does in fact expect that. It was assumed that it not returning the error state it got in turn from BakeBlendshape was the cause of this defect. However, it turns out that BakeBlendshape is also being called from somewhere else (as well as from BakeBlendshapeFromScript), and that is more likely the cause of this defect. Certainly using the repro steps from this defect we do not see an error in BakeBlendshapeFromScript at the time it is called from Python. But shortly after, so it is very likely related to this asset change but we've not quite found the direct link yet.

Call stack below for reference.. and we'll keep investigating.


>	_trinity_dx11_dev.dllTriGrannyRes::BakeBlendshape(unsigned int meshIx=0, const std::vector<float,std: - allocator<float> > & weights=[0](), void * pVertexData=0x35f27fd8, unsigned int vertexDataSize=840, const std::map<std::basic_string<char,std::char_traits<char>,std: - allocator<char> >,float,std::less<std::basic_string<char,std::char_traits<char>,std: - allocator<char> > >,std: - allocator<std::pair<std::basic_string<char,std::char_traits<char>,std: - allocator<char> > const ,float> > > * const nameToWeight=0x00000000 [...](), bool deltaOnly=true)  Line 396	C++
 	_trinity_dx11_dev.dllTriGrannyRes::GetBlendDeltas(unsigned int meshIx=0, const std::vector<float,std: - allocator<float> > & weights=[0](), std::vector<float,std: - allocator<float> > & deltaXyz=[210](0.00000000,,...,...))  Line 347	C++
 	_trinity_dx11_dev.dllTr2ClothingActor::AddToApexScene(Tr2ApexScene * apexScene=0x0aeaf288)  Line 173 + 0x27 bytes	C++
 	_trinity_dx11_dev.dllTr2ClothingActor::UpdateState(const Matrix & globalPose={...}, float * skinningMatrices=0x34f2c5f0, unsigned int skinningMatrixCount=119, Tr2ClothingActor::UpdateStateFlag continuousFlag=IS_CONTINUOUS, Tr2ApexScene * apexScene=0x0aeaf288)  Line 283	C++
 	_trinity_dx11_dev.dllTr2SkinnedObject::UpdateBones(__int64 time=130475584016449692, Tr2ApexScene * apexScene=0x0aeaf288)  Line 398	C++
 	_trinity_dx11_dev.dllTr2SkinnedObject::PostPhysicsUpdate(__int64 time=130475584016449692, Tr2ApexScene * apexScene=0x0aeaf288)  Line 146	C++
 	_trinity_dx11_dev.dllTr2IntSkinnedObject::PostPhysicsUpdate(__int64 time=130475584016449692, Tr2ApexScene * apexScene=0x0aeaf288)  Line 109	C++
 	_trinity_dx11_dev.dllTr2InteriorScene::Update(__int64 time=130475584016449692)  Line 715	C++
 	_trinity_dx11_dev.dllTriStepUpdate::Execute(__int64 time=130475584016449692, Tr2RenderContext & renderContext={...})  Line 23	C++
 	_trinity_dx11_dev.dllTriRenderJob::Run(__int64 time=130475584016449692, Tr2RenderContext * renderContextPtr=0x0447a588)  Line 97 + 0x3e bytes	C++
 	_trinity_dx11_dev.dllTr2RenderJobs::RunJobs(unsigned int affinity=0)  Line 176 + 0xc2 bytes	C++
 	_trinity_dx11_dev.dllTr2RenderJobs::Run(__int64 time=130475584016449692)  Line 212	C++
 	_trinity_dx11_dev.dllTriDevice::Render()  Line 1309	C++
 	_trinity_dx11_dev.dllTriDevice::HandleRenderTick(__int64 timestamp=130475584016449692)  Line 91 + 0x8 bytes	C++
 	_trinity_dx11_dev.dllTriDevice::OnTick(__int64 realTime=130475584016449692, __int64 simTime=130475584016449692, void * cookie=0x0671f9f4)  Line 791	C++

Review: http://codecollaborator/ui#review:id=3634

Binaries submitted in 801642
Build time : 9 minutes 27.328000 seconds
Branch : //eve/branches/staging/2014-DOLLY/

Binaries submitted in 801837
Build time : 9 minutes 21.276000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 801814]