DEFECT: Wyvern model does not show fully cloaked after gate jump https://jira.ccpgames.com/browse/EVE-117881
The cause of this is a wrong bounding sphere. You can clearly see the butt of the ship sticking out of it. Just another fail of our bounding sphere calculation algorithm. So I replaced it with a self written alogrithm based on the same system. So please expect to see tiny differences in the results.

Binaries submitted in 882639
Build time : 13 minutes 12.855000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 882632]