[eve MAINLINE] Automatic VCXPROJ submission from 1290
This might create build loops, but it was deemed valuable enough to be able to sync these.
https://teamcity.evetech.net/viewLog.html?buildId=5126865
!nobuild

Source Changes
----------------
[Eve Engine] 1466294 by patrick.kerr (Patrick Kerr)
cairo: Adding x64/v141 builds of the cairo and cairo-script DLLs.
----------------
[Eve Engine] 1466292 by patrick.kerr (Patrick Kerr)
cairo: Adding Win32/v141 builds of the cairo and cairo-script DLLs.
----------------
[Eve Engine] 1466291 by patrick.kerr (Patrick Kerr)
freetype 2.6.0: Update the "freetype" SDK for VS2017 and PlatformToolset v141.
----------------
[Eve Engine] 1466289 by patrick.kerr (Patrick Kerr)
libjpg-9b: Adding PlatformToolset "v141" support for the "libjpg" sdk (including explicit use of a "v100" folder for the older version, for consistency).
----------------
[Eve Engine] 1466287 by patrick.kerr (Patrick Kerr)
directx sdk: Conditionally skip the definition of various D3D constants which also exist in the newer Windows 10 SDK, as used by our VS2017 builds.
Many of these macros are now defined in "windows kits\10\include\10.0.14393.0\shared\winerror.h"
This isn't an ideal approach, because we probably shouldn't be including these old SDK files at all in the VS2017 builds (and we shouldn't be editing SDK files either!), but it's probably acceptable as a transitionary "hack".
The threshold is for skipping is _MSC_VER >= 1900, i.e. VS2015 or above.
(Oringinally unshelved from pending changelist '1399046' by kristjang).
----------------

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1466321]