Fix SystemError when calling ParseLabelText with valid arguments (#5)
This was using the deprecated "u" format specifier, which used to parse a Python string or Unicode object directly to a wide character string.

The "u" format specifier is no longer supported when calling PyArg_ParseTuple. As with any other unrecognized characters, passing "u" results in a SystemError. Therefore, use the "U" format specifier to parse a Python Unicode Object, then convert that to a wide string in order to get the expected wide character string.

PLAT-6013