Fixing a clipsphere issue
If the clipsphere center is outside of the boundingradius we get into a bad situation.

The bad situation is that we will then ALWAYS clip EVERYTHING if invert is set to true in the shader when we run this line:
	clip( invert * ( sign( clipRadius2Sq ) * dot( dist, dist ) - clipRadius2Sq ) );
because the first part of the subtraction will always be positive and therefore the clip will always happen.

The fix is to move the "nearDist" which is just the diff between clipradius and the boundingsphereradius into the normalized boundingradius, therefore making the clipRadii 0 when the clipspherefactors are 0.

https://ccpgames.atlassian.net/browse/PLAT-9093